refactor(agent-status): extract AgentStatusReconciliation namespace
Move the generic seededReducerState/hydratedStatus bridge helpers off
WorklaneStore into a shared @MainActor AgentStatusReconciliation enum so
the store and forthcoming per-tool status resolvers share one copy.
Verbatim move; all call sites retargeted.
